Basic algorithm implemented Test cases for input player data

+++ RiotGamesCodingChallenge/src/main/java/com/riotgames/interview/hongkong/matchmaking/matcher/MatchingAlgorithm.java
@@ -4,8 +4,8 @@ package com.riotgames.interview.hongkong.matchmaking.matcher;
  * Matching algorithm. Calculates a similarity score between two objects 
  * */
 public interface MatchingAlgorithm<T> {
-	
+
 	/** Similarity function. It should return a value between 0 (no match at all) and 1 (perfect match) */
-	public float getSimilarity(T one, T another);
+	public double getSimilarity(T one, T another);
 	
 }
